Cleanup `jupyter-request' and change its semantics
The purpose of this function was originally to define an action that
would generate a request when the state was a client and set up the
request so that it could be sent with the `jupyter-sent` action when
bound later on.  Instead, it now actually does the sending of the
request when bound making `jupyter-sent' superfluous.

* jupyter-monads.el (jupyter-subscribe-client): New function.
(jupyter-sent): Replace with `jupyter-do`.
(jupyter-request): Use `jupyter-mlet*`, send request as part of its
action. Add the option of passing in message subscribers as part of
the state.  If the state is a list with the first element being a
client, then the rest of the elements are message subscribers which
will be subscribed to the request's message publisher in the order
given.
(jupyter-message-subscribed): Account for the changes in
`jupyter-request'.
(jupyter-idle): Replace `jupyter-sent` with `jupyter-do`.

* jupyter-client.el (jupyter-interrupt-kernel)
(jupyter-handle-input-request, jupyter-eval-string)
(jupyter-completion-prefetch): Replace `jupyter-sent` with
`jupyter-do`.

* jupyter-julia.el (jupyter-julia--setup-hooks): Ditto.

* jupyter-org-client.el (jupyter-org-maybe-queued): Ditto.

* jupyter-repl.el (jupyter-repl-execute-cell)
(jupyter-eval-string, jupyter-repl-mode): Ditto.

* jupyter-widget-client.el (jupyter-widgets-on-message): Ditto.

* test/jupyter-test.el (jupyter-idle-sync)
(jupyter-zmq-channel-ioloop-send-fast): Ditto.

* test/test-helper.el (jupyter-test-with-client-cache): Ditto.