Killing a REPL buffer still allows one to be made later on
Previously, killing a REPL buffer only happened if not connected to a
kernel's communication channel or when a user wanted to shutdown a
kernel. Now, a REPL buffer is killed regardless of the connection
status or when a kernel is requested to also be shutdown.

* jupyter-repl.el (jupyter-repl-kill-buffer-query-function): Shutdown
a kernel if requested, disconnect from a kernel unconditionally,
always kill the REPL buffer unconditionally.
(jupyter-resurrect-repl): New function.
(jupyter-repl-pop-to-buffer): Use it.
(jupyter-bootstrap-repl): Ensure kernel communication is connected
first.